requirements.txt
annotated-types==0.7.0
antlr4-python3-runtime==4.11.1
asttokens==2.4.1
attrs==24.2.0
azure-core==1.32.0
azure-identity==1.17.1
backoff==2.2.1
beautifulsoup4==4.12.3
blinker==1.9.0
boto3==1.35.64
botocore==1.35.64
Brotli==1.1.0
certifi==2024.8.30
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
comm==0.2.2
contourpy==1.3.1
cramjam==2.9.0
cryptography==43.0.3
cycler==0.12.1
dacite==1.8.1
dash==2.18.2
dash-bootstrap-components==1.6.0
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
dash_colorscales==0.0.4
dash_daq==0.5.0
DataProfiler==0.12.0
datasketches==5.1.0
debugpy==1.8.7
decorator==5.1.1
deepdiff==6.7.1
Deprecated==1.2.14
dtale==3.15.1
duckdb==1.0.0
et_xmlfile==2.0.0
executing==2.1.0
fastavro==1.9.7
Flask==3.0.3
Flask-Compress==1.17
fonttools==4.55.0
future==1.0.0
googleapis-common-protos==1.65.0
h5py==3.12.1
HLL==2.2.0
htmlmin==0.1.12
ia-questionary==1.10.0
idna==3.10
ImageHash==4.3.1
importlib-metadata==6.11.0
importlib_resources==6.4.5
inflect==7.4.0
ipykernel==6.29.5
ipython==8.29.0
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.4
jmespath==1.0.1
joblib==1.4.2
jupyter_client==8.6.3
jupyter_core==5.7.2
kaleido==0.2.1
kiwisolver==1.4.7
llvmlite==0.43.0
lz4==4.3.3
markdown-it-py==3.0.0
MarkupSafe==2.1.2
matplotlib==3.9.2
matplotlib-inline==0.1.7
mdurl==0.1.2
missingno==0.5.2
more-itertools==10.5.0
msal==1.31.1
msal-extensions==1.2.0
multimethod==1.12
nest-asyncio==1.6.0
networkx==3.4.2
numba==0.60.0
numpy==1.26.4
openpyxl==3.1.5
opentelemetry-api==1.22.0
opentelemetry-exporter-otlp-proto-common==1.22.0
opentelemetry-exporter-otlp-proto-http==1.22.0
opentelemetry-proto==1.22.0
opentelemetry-sdk==1.22.0
opentelemetry-semantic-conventions==0.43b0
ordered-set==4.1.0
packaging==24.1
pandas==2.2.3
parso==0.8.4
patsy==1.0.1
phik==0.12.4
pillow==11.0.0
platformdirs==4.3.6
plotly==5.24.1
portalocker==2.10.1
prompt-toolkit==3.0.36
protobuf==4.25.5
psutil==6.1.0
pure_eval==0.2.3
pyarrow==18.0.0
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
Pygments==2.18.0
PyJWT==2.9.0
pyodbc==5.2.0
pyparsing==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-snappy==0.7.3
pytz==2024.2
PyWavelets==1.7.0
pywin32==308
PyYAML==6.0.2
pyzmq==26.2.0
requests==2.32.3
retrying==1.3.4
rich==13.9.3
ruamel.yaml==0.17.40
ruamel.yaml.clib==0.2.12
s3transfer==0.10.3
scikit-learn==1.5.2
scipy==1.13.1
seaborn==0.13.2
segment-analytics-python==2.3.3
setuptools==75.2.0
shellingham==1.5.4
six==1.16.0
soda==1.7.1
soda-core==3.4.1
soda-core-sqlserver==3.4.1
soda-duckdb==1.7.1
soda-sqlserver==1.7.1
soda_freshness==0.1.1
soupsieve==2.6
sqlparse==0.5.1
squarify==0.4.4
stack-data==0.6.3
statsmodels==0.14.4
strsimpy==0.2.1
sweetviz==2.3.1
tenacity==9.0.0
threadpoolctl==3.5.0
tornado==6.4.1
tqdm==4.67.0
traitlets==5.14.3
typeguard==4.4.0
typer==0.12.5
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
visions==0.7.6
wcwidth==0.2.13
Werkzeug==3.0.6
wheel==0.45.0
wordcloud==1.9.4
wrapt==1.16.0
xarray==2024.10.0
xlrd==2.0.1
ydata-profiling==4.12.0
zipp==3.20.2
zstandard==0.23.0
